Add API key enabled field
Conjur will use this field to determine whether or not the given rule should have an API key generated for it.

Some roles are not expected to use default API key authentication, but rather are expected to only use a given authenticator, such as `authn-k8s`. When this is the case, the API key becomes a backdoor to allow impersonating this role. This allows a Conjur operator to explicitly disable the API key, preventing this behavior.